Security enhanced, controllers were provided with necessary checks when the logged user wants to modify the content.

+++ src/main/java/app/controllers/RateController.java
@@ -1,12 +1,16 @@
 package app.controllers;
+import app.Exceptions.ActionNotAuthorizedException;
+import app.Role;
 import app.dtos.creating.CreateRateDTO;
 import app.dtos.responding.ResponseRateDTO;
 import app.dtos.updating.UpdateRateDTO;
 import app.models.Rate;
+import app.security.MyUserDetails;
 import app.services.RateService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -23,12 +27,32 @@ public class RateController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteRateById(@PathVariable Long id,  @PathVariable Long cafeId) {
+        MyUserDetails currentUser = (MyUserDetails) SecurityContextHolder
+                .getContext()
+                .getAuthentication()
+                .getPrincipal();
+
+
+
+        if (!currentUser.getId().equals(id) && !currentUser.getAuthorities().contains(Role.ADMIN)) {
+            throw new ActionNotAuthorizedException("You are not authorized to delete this rate");
+        }
         rateService.deleteRateById(cafeId, id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<ResponseRateDTO> updateRateById(@PathVariable Long id, @PathVariable Long cafeId, @RequestBody UpdateRateDTO dto) {
+        MyUserDetails currentUser = (MyUserDetails) SecurityContextHolder
+                .getContext()
+                .getAuthentication()
+                .getPrincipal();
+
+
+
+        if (!currentUser.getId().equals(id) && !currentUser.getAuthorities().contains(Role.ADMIN)) {
+            throw new ActionNotAuthorizedException("You are not authorized to delete this rate");
+        }
         rateService.updateRate(dto, cafeId, id);
         ResponseRateDTO rateDTO =  rateService.getRateDtoById(cafeId, id);
         return new ResponseEntity<>(rateDTO, HttpStatus.OK);
